ai_backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
mediapipe
numpy
sqlalchemy
websockets
moviepy==1.0.3
pillow
